Making ObstacleMapManager(shouldUpdate,receiveMap), Robot and Server : send and receive map function. Chenging MapKey: Long -> Int

+++ src/main/scala/edu/hkcc/pacmanrobot/utils/map/MapKey.scala
@@ -3,7 +3,7 @@ package edu.hkcc.pacmanrobot.utils.map
 /**
  * Created by beenotung on 3/21/15.
  */
-class MapKey(x: Long, y: Long) extends Cloneable {
+class MapKey(x: Int, y: Int) extends Cloneable {
   override def clone: AnyRef = {
     new MapKey(x, y)
   }
